feat(schema): Use 'string' for bigserial and biginteger
PostgreSQL's 'bigserial' and 'bigint' types are 64-bit. This can't be expressed with JavaScript's
'Number' type, so a 'string' type with a pattern should be used instead.

BREAKING CHANGE: Columns that are of 'bigint' or 'bigserial' types will now be described in the JSON Schema
document as 'string' instead of 'number'.