feat: Upgrade JSON Schema version from `draft-04` to `draft-07`
Strictly speaking, it was possible for previous version of Collimator
to generate a Schema that *was not* valid according to v4 semantics, as
it was possible to end up with a `required` array that was empty, which
is invalid according to the v4 spec.

An empty array was made permissible under the v6 spec, and at the time
of writing, v7 is the latest, so it makes sense to use the latest
version.

BREAKING CHANGE: The Schema document generated for inspected tables is
now versioned as `"http://json-schema.org/draft-07/schema#"`.